% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.namespace.r
\name{add.namespace}
\alias{add.namespace}
\title{Add query namespaces}
\usage{
add.namespace(query)
}
\arguments{
\item{query}{A SPARQL query string that you want namespaces for}
}
\value{
Query string with namespaces
}
\description{
Add query namespaces
}
\keyword{internal}

