% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaProdByInd.R
\name{beaProdByInd}
\alias{beaProdByInd}
\title{Pass list of user specifications (including API key) to return data from BEA API.}
\usage{
beaProdByInd(beaKey, component, geofips = "State", indVec = c())
}
\arguments{
\item{beaKey}{Character string representation of user's 36-digit BEA API key}

\item{component}{Character string representation of component parameter (see API documentation)}

\item{geofips}{Geographic FIPS code (see API documentation)}

\item{indVec}{Character vector containing industryIDs (typically obtained from "shortlist()" function}
}
\value{
By default, an object of class 'data.table' containing BEA product data by industry and region
}
\description{
Pass list of user specifications (including API key) to return data from BEA API.
}
\keyword{internal}

