% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blank_ID.r
\name{Blank_ID}
\alias{Blank_ID}
\title{Utility function to create a blank node ID}
\usage{
Blank_ID(blank.counter)
}
\arguments{
\item{blank.counter}{counter distinguishing the blank nodes}
}
\value{
a blank node ID
}
\description{
Utility function to create a blank node ID
}

