% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeList_To_RDF.r
\name{CodeList_To_RDF}
\alias{CodeList_To_RDF}
\title{Inserts an SDMX codelist into a triple store as an RDF SKOS ConceptScheme.}
\usage{
CodeList_To_RDF(SPARQL.endpoint, code.lists)
}
\arguments{
\item{SPARQL.endpoint}{the SPARQL endpoint URL of the target triple store}

\item{code.lists}{a list of SDMX codelist objects}
}
\description{
Inserts an SDMX codelist into a triple store as an RDF SKOS ConceptScheme.
}

