% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataflow_To_RDF.r
\name{Dataflow_To_RDF}
\alias{Dataflow_To_RDF}
\title{Inserts an SDMX dataflow into a triple store as an RDF Data Cube Vocabulary dataset. The dataset is complemented with metadata which follows the DCAT ontology.}
\usage{
Dataflow_To_RDF(SPARQL.endpoint, dataflow)
}
\arguments{
\item{SPARQL.endpoint}{the SPARQL endpoint URL of the target triple store}

\item{dataflow}{SDMX dataflow object}
}
\description{
Inserts an SDMX dataflow into a triple store as an RDF Data Cube Vocabulary dataset. The dataset is complemented with metadata which follows the DCAT ontology.
}

