% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeRel.R
\name{describeRel}
\alias{describeRel}
\title{Given a relationship ID, return description of mapped relationship allowing users to see common overlap and/or differences between datasets}
\usage{
describeRel(term, beaKey = "")
}
\arguments{
\item{term}{ID of the statistic requested}

\item{beaKey}{BEA API key (won't be necessary once SPARQL repository has been updated with timestamp)}
}
\description{
Given a relationship ID, return description of mapped relationship allowing users to see common overlap and/or differences between datasets
}

