% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_to_RDF.r
\name{DSD_To_RDF}
\alias{DSD_To_RDF}
\title{Inserts an SDMX DSD into a triple store as an RDF Data Cube Vocabulary DataStructureDefinition.}
\usage{
DSD_To_RDF(SPARQL.endpoint, data.structure)
}
\arguments{
\item{SPARQL.endpoint}{the SPARQL endpoint URL of the target triple store}

\item{data.structure}{SDMX DSD object}
}
\description{
Inserts an SDMX DSD into a triple store as an RDF Data Cube Vocabulary DataStructureDefinition.
}

