% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDetect.R
\name{geoDetect}
\alias{geoDetect}
\title{detect, extract, and merge shapefiles}
\usage{
geoDetect(data)
}
\arguments{
\item{data}{data.table object from getRel()}
}
\value{
Shapefile of relevant geographic units
}
\description{
detect, extract, and merge shapefiles
}
\keyword{internal}

