% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoMap.R
\name{geoMap}
\alias{geoMap}
\title{create interactive map of dataset for a given year}
\usage{
geoMap(dataset, year, asSHP = FALSE,
  tile_provider = "http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  fillOpacity = 0.8, colorPal = "Blues", colorSteps = 30)
}
\arguments{
\item{year}{selected year for map, "all" returns a shapefile with all years of data in wide form}

\item{asSHP}{controls output as a leaflet map or a merged SHP}

\item{data}{data.table object from getRel()}
}
\value{
Shapefile of relevant geographic units
}
\description{
create interactive map of dataset for a given year
}
\keyword{internal}

