% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStat.r
\name{getStat}
\alias{getStat}
\title{Return data from the eurostat and bea APIs}
\usage{
getStat(statID = "", NUTS = "2", euParms = list(), usParms = list(),
  apiKey = NULL, meta = FALSE, freq = "A")
}
\arguments{
\item{statID}{ID of the statistic requested}
}
\value{
By default, an object of class 'data.table'
}
\description{
Return data from the eurostat and bea APIs
}

