% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large.query.handler.r
\name{large.query.handler}
\alias{large.query.handler}
\title{This function takes a list of RDF statements (triples) and: 1) divides them into chunks; 2) turns them into a full-fledged SPARQL INSERT query using the query.builder function; and 3) inserts them into the target triple store by executing the SPARQL INSERT query}
\usage{
large.query.handler(endpoint, query.list, query.builder, limit = 5000)
}
\arguments{
\item{endpoint}{the SPARQL endpoint URL of the target triple store}

\item{query.list}{a list of RDF statements}

\item{query.builder}{the function to be used to build the query}

\item{limit}{maximum number of characters in each chunk}
}
\description{
This function takes a list of RDF statements (triples) and: 1) divides them into chunks; 2) turns them into a full-fledged SPARQL INSERT query using the query.builder function; and 3) inserts them into the target triple store by executing the SPARQL INSERT query
}

