% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listRel.R
\name{listRel}
\alias{listRel}
\title{List all datasets available}
\usage{
listRel(asHtml = TRUE)
}
\arguments{
\item{asHtml}{Option to render results in an interactive DT}
}
\description{
List all datasets available
}

