% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLocalAcronyms.r
\name{loadLocalAcronyms}
\alias{loadLocalAcronyms}
\title{Load local acronym metadata cache}
\usage{
loadLocalAcronyms()
}
\value{
Local cache of acronym metadata
}
\description{
Load local acronym metadata cache
}
\keyword{internal}

