% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLocalRel.r
\name{loadLocalRel}
\alias{loadLocalRel}
\title{Load local relationship metadata cache}
\usage{
loadLocalRel()
}
\value{
Local cache of relationship metadata, as data.table
}
\description{
Load local relationship metadata cache
}
\keyword{internal}

