% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propertySub.R
\name{propertySub}
\alias{propertySub}
\title{Quick query to Dydra repo for properties to be used as labels and substitution from filtered localMerge in getRel}
\usage{
propertySub(localMrg = loadLocalMerge(), filterIDs = c())
}
\arguments{
\item{filterIDs}{Character vector of IDs to filter property map, typically passed by getRel}

\item{localMerge}{Local Merge metadata, typically passed by getRel}
}
\value{
Value map to be used in getRel
}
\description{
Quick query to Dydra repo for properties to be used as labels and substitution from filtered localMerge in getRel
}
\keyword{internal}

