% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.builder.r
\name{query.builder}
\alias{query.builder}
\title{Creates a well-formed INSERT SPARQL query by adding the INSERT clause and the required namespace prefixes}
\usage{
query.builder(query)
}
\arguments{
\item{query}{the desired content of the INSERT clause}
}
\value{
A well-formed SPARQL INSERT query
}
\description{
Creates a well-formed INSERT SPARQL query by adding the INSERT clause and the required namespace prefixes
}

