% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchParse.R
\name{searchParse}
\alias{searchParse}
\title{Parses, spell checks, and prepares a search query for standardization
by creating n-grams}
\usage{
searchParse(rec, spellCheck = TRUE)
}
\arguments{
\item{rec}{Character string for which to recommend terms}

\item{spellCheck}{Is true to force spellcheck}
}
\description{
Parses, spell checks, and prepares a search query for standardization
by creating n-grams
}

