% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchRefine.R
\name{searchRefine}
\alias{searchRefine}
\title{Refines search query by spellchecking and standardizing acronyms in query string}
\usage{
searchRefine(rec)
}
\arguments{
\item{rec}{is a character string with the terms to be searched}
}
\description{
Refines search query by spellchecking and standardizing acronyms in query string
}

