% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statSearch.r
\name{statSearch}
\alias{statSearch}
\title{Search metadata from the eurostat and bea APIs}
\usage{
statSearch(searchTerm, apiKey = NULL)
}
\arguments{
\item{statID}{ID of the statistic requested}
}
\value{
By default, an object of class 'data.table'
}
\description{
Search metadata from the eurostat and bea APIs
}

