% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSync.R
\name{timeSync}
\alias{timeSync}
\title{Provides user easier access to extracts the latest overlapping year of data or all overlapping years.}
\usage{
timeSync(dataset, sync)
}
\arguments{
\item{dataset}{data.frame or data.table}

\item{sync}{Numeric value representing period, with options: 1 = latest period, 2 = overlapping period}
}
\description{
Provides user easier access to extracts the latest overlapping year of data or all overlapping years.
}

