% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLocalMerge.r
\name{updateLocalMerge}
\alias{updateLocalMerge}
\title{Update local merge metadata cache}
\usage{
updateLocalMerge()
}
\value{
Local cache of merge metadata, as data.table
}
\description{
Update local merge metadata cache
}
\keyword{internal}

