% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLocalRel.r
\name{updateLocalRel}
\alias{updateLocalRel}
\title{Update local relationship metadata cache}
\usage{
updateLocalRel()
}
\value{
Local cache of relationship metadata, as data.table
}
\description{
Update local relationship metadata cache
}
\keyword{internal}

