% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLocalAcronyms.R
\name{loadLocalAcronyms}
\alias{loadLocalAcronyms}
\title{Load local acronym metadata cache}
\usage{
loadLocalAcronyms()
}
\value{
Local cache of acronyms
}
\description{
Load local acronym metadata cache
}
\keyword{internal}

