% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCache.R
\name{updateCache}
\alias{updateCache}
\title{Update local metadata cache}
\usage{
updateCache(beaKey, force = FALSE)
}
\arguments{
\item{beaKey}{BEA API key (won't be necessary once SPARQL repository has been updated with timestamp)}

\item{force}{Force local cache to update, even if there have been no changes since your last metadata update}
}
\value{
Nothing, but updates local CSVs
}
\description{
Update local metadata cache
}

