% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCache.R
\name{updateCache}
\alias{updateCache}
\title{Update local metadata cache}
\usage{
updateCache(beaKey)
}
\arguments{
\item{beaKey}{BEA API key (won't be necessary once SPARQL repository has been updated with timestamp)}
}
\value{
Nothing, but updates local CSVs
}
\description{
Update local metadata cache
}

